Fix Square payment processing to correctly handle monetary amounts
Update server/services/squareService.ts to ensure Square payment amounts are processed as regular numbers in cents, resolving payment failures.

+++ server/services/squareService.ts
@@ -10,7 +10,7 @@ class SquareService {
     return {
       id: `sq_payment_${Date.now()}`,
       status: "COMPLETED",
-      amount: BigInt(Math.round(amount * 100)),
+      amount: Math.round(amount * 100), // Convert to cents as regular number
       currency: "USD",
     };
   }
